+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [Unreleased]
+### Improved
+- it has been made that the shipments registration takes place via cronjob, when the shipments is registered via bulk action on the Woocommerce order page
+
 ## [1.6.0] - 2025-06-09
 ### Fixed
 - disabled test parameter